+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_to_dict_v3.py
@@ -54,7 +54,7 @@ class TestToDict(FormRecognizerTest):
         assert d == final
 
     def test_document_element_to_dict(self):
-        model = _models.DocumentElement(
+        model = _models.DocumentContentElement(
             content="sample",
             bounding_box=[
                 _models.Point(1427.0, 1669.0),
@@ -844,6 +844,57 @@ class TestToDict(FormRecognizerTest):
             column_index=3,
             row_span=1,
             column_span=1,
+            content="cell content",
+            bounding_regions=[
+                _models.BoundingRegion(
+                    bounding_box=[_models.Point(1, 2), _models.Point(3, 4)],
+                    page_number=1,
+                ),
+            ],
+            spans=[
+                _models.DocumentSpan(
+                    offset=5,
+                    length=2,
+                ),
+            ],
+        )
+
+        d = model.to_dict()
+
+        final = {
+            "kind": "content",
+            "row_index": 2,
+            "column_index": 3,
+            "row_span": 1,
+            "column_span": 1,
+            "content": "cell content",
+            "bounding_regions": [
+                {
+                    "page_number": 1,
+                    "bounding_box": [
+                        {"x": 1, "y": 2},
+                        {
+                            "x": 3,
+                            "y": 4,
+                        },
+                    ],
+                },
+            ],
+            "spans": [
+                {
+                    "offset": 5,
+                    "length": 2,
+                },
+            ],
+        }
+
+        assert d == final
+
+    def test_document_table_cell_to_dict_use_defaults(self):
+        # NOTE: kind, column_span, and row_span are not included on purpose to test that the proper defaults are set.
+        model = _models.DocumentTableCell(
+            row_index=2,
+            column_index=3,
             content="cell content",
             bounding_regions=[
                 _models.BoundingRegion(