+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/_file_system_client.py
@@ -4,8 +4,7 @@
 # license information.
 # --------------------------------------------------------------------------
 import functools
-from typing import Optional, Any, Union, Iterator
-
+from typing import Optional, Any, Union, TypeVar, Iterator
 
 try:
     from urllib.parse import urlparse, quote, unquote
@@ -21,7 +20,7 @@ from azure.core.paging import ItemPaged
 from azure.storage.blob import ContainerClient
 from ._shared.base_client import TransportWrapper, StorageAccountHostsMixin, parse_query, parse_connection_str
 from ._serialize import convert_dfs_url_to_blob_url, get_api_version
-from ._list_paths_helper import DeletedPathPropertiesPaged
+from ._list_paths_helper import DeletedPathPropertiesPaged, PathPropertiesPaged
 from ._models import LocationMode, FileSystemProperties, PublicAccess, DeletedPathProperties, FileProperties, \
     DirectoryProperties
 from ._data_lake_file_client import DataLakeFileClient
@@ -29,7 +28,10 @@ from ._data_lake_directory_client import DataLakeDirectoryClient
 from ._data_lake_lease import DataLakeLeaseClient
 from ._generated import AzureDataLakeStorageRESTAPI
 from ._generated.models import ListBlobsIncludeItem
-from ._deserialize import deserialize_path_properties, process_storage_error, is_file_path
+from ._deserialize import process_storage_error, is_file_path
+
+
+ClassType = TypeVar("ClassType")
 
 
 class FileSystemClient(StorageAccountHostsMixin):
@@ -137,11 +139,12 @@ class FileSystemClient(StorageAccountHostsMixin):
 
     @classmethod
     def from_connection_string(
-            cls, conn_str,  # type: str
+            cls,  # type: Type[ClassType]
+            conn_str,  # type: str
             file_system_name,  # type: str
             credential=None,  # type: Optional[Any]
             **kwargs  # type: Any
-        ):  # type: (...) -> FileSystemClient
+        ):  # type: (...) -> ClassType
         """
         Create FileSystemClient from a Connection String.
 
@@ -513,13 +516,14 @@ class FileSystemClient(StorageAccountHostsMixin):
                 :caption: List the paths in the file system.
         """
         timeout = kwargs.pop('timeout', None)
-        return self._client.file_system.list_paths(
-            recursive=recursive,
-            max_results=max_results,
+        command = functools.partial(
+            self._client.file_system.list_paths,
             path=path,
             timeout=timeout,
-            cls=deserialize_path_properties,
             **kwargs)
+        return ItemPaged(
+            command, recursive, path=path, max_results=max_results,
+            page_iterator_class=PathPropertiesPaged, **kwargs)
 
     def create_directory(self, directory,  # type: Union[DirectoryProperties, str]
                          metadata=None,  # type: Optional[Dict[str, str]]