+++ sdk/core/azure-core/CHANGELOG.md
@@ -10,8 +10,11 @@
 
 - respect text encoding specified in argument (thanks to @ryohji for the contribution)  #20796
 - Fix "coroutine x.read() was never awaited" warning from `ContentDecodePolicy`  #21318
+- fix type check for `data` input to `azure.core.rest` for python 2.7 users  #21341
 
 ### Other Changes
+
+- Refactor AzureJSONEncoder (thanks to @Codejune for the contribution)  #21028
 
 ## 1.19.0 (2021-09-30)