+++ sdk/purview/azure-purview-administration/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.0.0b2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.0.0b1 (2021-09-29)
 
 - This is the initial release of the Azure Purview Administration library.